[ottf] Completely revamp the OTTF console subsystem
