+++ sw/device/lib/testing/test_framework/BUILD
@@ -305,9 +305,16 @@ cc_library(
 dual_cc_library(
     name = "ottf_console",
     srcs = dual_inputs(
-        device = ["ottf_console.c"],
+        device = [
+            "ottf_console.c",
+            "ottf_console_uart.c",
+            "ottf_console_spi.c",
+        ],
     ),
-    hdrs = ["ottf_console.h"],
+    hdrs = [
+        "ottf_console.h",
+        "ottf_console_internal.h",
+    ],
     deps = dual_inputs(
         device = [
             ":check",
