+++ sw/device/lib/testing/test_framework/ottf_test_config.h
@@ -13,7 +13,7 @@ typedef enum ottf_console_type {
   kOttfConsoleSpiDevice,
 } ottf_console_type_t;
 
-typedef struct ottf_console {
+typedef struct ottf_console_t {
   /**
    * Communication interface type to use for the OTTF console (see
    * `ottf_console_type_t` above).
@@ -38,7 +38,7 @@ typedef struct ottf_console {
    * transmissions.
    */
   bool putbuf_buffered;
-} ottf_console_t;
+} ottf_console_opt_t;
 
 typedef struct ottf_console_tx_indicator {
   /**
@@ -91,7 +91,7 @@ typedef struct ottf_test_config {
    * status and error messages are written to. Typically UART0, but other
    * communication peripherals may be supported.
    */
-  ottf_console_t console;
+  ottf_console_opt_t console;
 
   /**
    * The TX indicator GPIO to use in conjunction with the SPI console, if a
